+++ src/main/java/fr/insee/rmes/bauhaus_services/rdf_utils/RdfUtils.java
@@ -143,10 +143,6 @@ public class RdfUtils {
 		return objectIRI(ObjectType.CODE_LIST, id);
 	}
 
-	public static IRI codesListIRI(String StringUri) {
-		return factory.createIRI(StringUri);
-	}
-
 
 	public static IRI createIRI(String uri){
 		return factory.createIRI(uri);
